update chat scrolling, add thinking indicator
