+++ src/components/chat/ThinkingMessage.svelte
@@ -1,0 +1,7 @@
+<div
+	class="w-4/5 h-10 flex items-end bg-blue-primary text-light p-2 pl-6 text-sm rounded-md space-x-2"
+>
+	<div class="animate-bounce">&bull;</div>
+	<div class="animate-bounce" style:animation-delay="100ms">&bull;</div>
+	<div class="animate-bounce" style:animation-delay="200ms">&bull;</div>
+</div>
